# perf(scan): improve speed of interpretation pipeline