--- a/libs/ballot-interpreter/src/types.ts
+++ b/libs/ballot-interpreter/src/types.ts
@@ -19,4 +19,6 @@ export interface InterpreterOptions {
   testMode: boolean;
   disableBmdBallotScanning?: boolean;
   minimumDetectedScale?: number;
+  frontNormalizedImageOutputPath?: string;
+  backNormalizedImageOutputPath?: string;
 }