… apps/mark-scan/backend/src/app.diagnostics.test.ts
… apps/mark-scan/backend/src/custom-paper-handler/state_machine.test.ts
… apps/mark-scan/backend/test/ballot_helpers.ts
… libs/ballot-interpreter/src/hmpb-ts/debug.test.ts
… libs/ballot-interpreter/src/hmpb-ts/interpret.test.ts
… libs/ballot-interpreter/src/interpret_all_bubble_ballot.test.ts
… libs/ballot-interpreter/src/interpret_and_save_files.test.ts
… libs/ballot-interpreter/src/interpret_blank_sheet.test.ts
… libs/ballot-interpreter/src/interpret_bmd_and_hmp_ballots.test.ts
… libs/ballot-interpreter/src/interpret_bmd_ballots.test.ts
… libs/ballot-interpreter/src/interpret_vx_hmpb.test.ts
… libs/ballot-interpreter/test/helpers/interpretation.ts