## Overview

This change provides a small speed boost to HMPB interpretation by changing the contract of the Rust code to never pass image data back, instead writing the normalized ballot image data to disk in the location that the calling JS code wants it to go.

## Demo Video or Screenshot
<details>
<summary>
Built both versions (JS writes, Rust writes) and ran them several times. Typical times on my i5 machine were <b>75ms</b> for Rust, <b>280ms</b> for JS.
</summary>
<pre>
❯ node compare.js
setup: 131.41ms
JS Writes: 283.72ms
Rust Writes: 72.745ms

❯ node compare.js
setup: 140.669ms
JS Writes: 296.398ms
Rust Writes: 79.841ms

❯ node compare.js
setup: 132.113ms
JS Writes: 274.477ms
Rust Writes: 75.326ms

❯ node compare.js
setup: 145.867ms
JS Writes: 303.028ms
Rust Writes: 81.836ms

❯ node compare.js
setup: 146.944ms
JS Writes: 288.04ms
Rust Writes: 77.003ms
</pre>
</details>

I did some further testing in the context of the VxScan app on my machine (middling Lenovo ThinkCentre i5) and got these numbers:

JS Writes:
- start scan: 0.70s
- end scan: 1.50s
- start eject: 2.04s
- hold time: 2.04s - 1.50s = 540ms

Rust Writes:
- start scan: 0.89s
- end scan: 1.71s
- start eject: 2.02s
- hold time: 2.02s - 1.71s = 310ms

So it seems like the timing using the comparison script roughly translates to the context of the VxScan app, saving about 200ms on this hardware.

## Testing Plan
<details>
<summary>
Tested performance using a script that loaded both versions of the Rust addon and either deferred to Rust ("Rust Writes") or handled writing itself ("JS Writes").
</summary>
<pre>
const jsWrites = require('./build/old.node');
const rustWrites = require('./build/new.node');
const { loadImageData } = require('@votingworks/image-utils');
const { saveSheetImages } = require('./build/save_images');
const assert = require('node:assert');

async function main() {
  console.time('setup');
  const election = require('./test/fixtures/104h-2025-04/election.json');
  const frontImageData = await loadImageData('./test/fixtures/104h-2025-04/imprinter-front.png');
  const backImageData = await loadImageData('./test/fixtures/104h-2025-04/imprinter-back.png');
  console.timeEnd('setup');

  console.time('JS Writes');
  const jsWritesResult = jsWrites.interpret(election, frontImageData, backImageData);
  assert(jsWritesResult.success);
  await saveSheetImages({
    sheetId: 'js-writes',
    ballotImagesPath: '.',
    images: [
      { ...jsWritesResult.frontNormalizedImage, data: new Uint8ClampedArray(jsWritesResult.frontNormalizedImage.data) },
      { ...jsWritesResult.backNormalizedImage, data: new Uint8ClampedArray(jsWritesResult.backNormalizedImage.data) },
    ],
  });
  console.timeEnd('JS Writes');

  console.time('Rust Writes');
  const rustWritesResult = rustWrites.interpret(election, frontImageData, backImageData, {
    frontNormalizedImageOutputPath: 'rust-writes-front.png',
    backNormalizedImageOutputPath: 'rust-writes-back.png',
  });
  console.timeEnd('Rust Writes');
}

main().catch((e) => console.error(e.stack));
</pre>
</details>